refactor(v0.5.2): bootstrap_workflow_kit.py → bootstrap_lib/ 6-module package
- Split 2,468-line monolithic bootstrap_workflow_kit.py into 6 focused modules
- bootstrap_lib/paths.py (158 lines) - Paths, HarnessDefinition, make_paths, 13 path helpers, IGNORED_DIRS
- bootstrap_lib/writes.py (154 lines) - write_text, copy_core_docs, build_manifest, rel
- bootstrap_lib/discovery.py (163 lines) - iter_repo_files, detect_package_scripts, guess_run_command, value_or_inferred, global_snippet_sources, IGNORED_DIRS
- bootstrap_lib/renderers.py (351 lines) - render_readme, render_project_profile, render_session_handoff, render_backlog_index, render_daily_backlog, render_assessment, render_project_status_assessment, load_template
- bootstrap_lib/mcp.py (255 lines) - MCP_SERVER_ALIAS, MCP_TOOL_NAME, _mcp_server_command, _mcp_server_env, 5 render_*_mcp_config, write_mcp_config_files, MCP_CONFIG_RENDERERS
- bootstrap_lib/__main__.py (831 lines) - parse_args, infer_project_context, selected_harnesses, update_dependencies, write_harness_files (dispatcher), main
- Move bootstrap_harnesses/ → bootstrap_lib/harnesses/ (registry preserved)
- bootstrap_lib/harnesses/renderers.py (1,046 lines) - 6 harness render_*_agents + 6 write_*_harness_files
- bootstrap_workflow_kit.py → 53-line thin re-export entry (backward compat)
- Memory layer: release/v0.5.2/ (PROJECT_PROFILE, session_handoff, backlog, state.json)

Regression: check_bootstrap.py 7/7 PASS (new/existing/opencode/gemini-cli/antigravity/minimax-code/enable_mcp), check_bootstrap_mcp_roundtrip.py 5/5 PASS, linter status ok (0 issues, 0 broken links)

TASK-V052-001 done
