+++ ai-workflow/memory/release/v0.5.2/backlog/2026-06-06.md
@@ -1,0 +1,87 @@
+# 2026-06-06 작업 백로그 (release/v0.5.2)
+
+- 문서 목적: 2026-06-06 세션에서 release/v0.5.2 브랜치에서 처리한 작업을 기록한다.
+- 범위: TASK-V052-001 (bootstrap 풀 리팩터), TASK-V052-002 (workflow_kit 패키지화), TASK-V052-003 (외부 pilot validation)
+- 대상 독자: AI 에이전트, 저장소 maintainer
+- 상태: in_progress
+- 최종 수정일: 2026-06-06
+- 관련 문서: [`../session_handoff.md`](../session_handoff.md), [`../../../work_backlog.md`](../../../work_backlog.md), [`../PROJECT_PROFILE.md`](../PROJECT_PROFILE.md)
+
+## TASK-V052-001 `bootstrap_workflow_kit.py` 풀 리팩터
+
+- 상태: done
+- 우선순위: high
+- 요청일: 2026-06-06
+- 완료일: 2026-06-06
+- 담당: yklee (Mavis / MiniMax Code orchestrator)
+- 호스트명: yklee-ui-MacBookAir.local
+- 호스트 IP: 192.168.0.139
+- 영향 문서: `workflow-source/scripts/bootstrap_workflow_kit.py` (53줄 thin entry), 신규 `workflow-source/scripts/bootstrap_lib/` (6 모듈)
+
+### 작업 내용
+
+- 2,468줄 모놀리식 `bootstrap_workflow_kit.py` → `bootstrap_lib/` 6 모듈 패키지화
+- v0.5.0 에서 abort 했던 작업을 "thin entry + 6 모듈" 로 단순화하여 재시도
+- `bootstrap_harnesses/` (154줄) → `bootstrap_lib/harnesses/` 이동 (기존 registry 그대로)
+- backward compat: `bootstrap_workflow_kit.py` 를 thin re-export entry 로 유지
+
+### 진행 현황
+
+- 2026-06-06 21:08 시작
+- 2026-06-06 22:00 `bootstrap_lib/` 6 모듈 골격 완성
+- 2026-06-06 22:30 thin entry 작성 + 회귀 확인
+- 2026-06-06 22:45 check_bootstrap.py 7/7 PASS (new/existing/opencode/gemini-cli/antigravity/minimax-code/enable_mcp)
+- 2026-06-06 22:50 check_bootstrap_mcp_roundtrip.py 5/5 PASS (5개 하네스 모두 round-trip)
+
+### 완료 기준
+
+- [x] `bootstrap_workflow_kit.py` 가 thin re-export entry (≤100줄)
+- [x] `bootstrap_lib/` 패키지 6 모듈 (paths/writes/discovery/renderers/mcp + harnesses/)
+- [x] `bootstrap_lib/harnesses/` 가 기존 registry 그대로 보존
+- [x] `check_bootstrap.py` 7/7 회귀 0
+- [x] `check_bootstrap_mcp_roundtrip.py` 5/5 회귀 0
+- [x] backward compat: `python bootstrap_workflow_kit.py` 직접 실행 가능
+- [x] memory layer 갱신 (handoff, backlog, state.json, linter)
+
+### 작업 결과
+
+- 신규 파일:
+  - `bootstrap_lib/__init__.py` (empty, 패키지 마커)
+  - `bootstrap_lib/paths.py` (158줄) - Paths, HarnessDefinition, make_paths, 13개 *_path helpers, IGNORED_DIRS
+  - `bootstrap_lib/writes.py` (154줄) - write_text, copy_core_docs, build_manifest, rel
+  - `bootstrap_lib/discovery.py` (163줄) - iter_repo_files, detect_package_scripts, guess_run_command, value_or_inferred, global_snippet_sources, IGNORED_DIRS, COMMON_*_DEPS
+  - `bootstrap_lib/renderers.py` (351줄) - render_readme, render_project_profile, render_session_handoff, render_backlog_index, render_daily_backlog, render_assessment, render_project_status_assessment, load_template, SOURCE_ROOT
+  - `bootstrap_lib/mcp.py` (255줄) - MCP_SERVER_ALIAS, MCP_TOOL_NAME, _mcp_server_command, _mcp_server_env, 5개 render_*_mcp_config, write_mcp_config_files, MCP_CONFIG_RENDERERS
+  - `bootstrap_lib/__main__.py` (831줄) - parse_args, infer_project_context, selected_harnesses, update_dependencies, write_harness_files (dispatcher), main, build_manifest, harness builder registration
+  - `bootstrap_lib/harnesses/__init__.py` (154줄, 이동) - HARNESS_SPECS, HARNESS_FILE_BUILDERS, SUPPORTED_HARNESSES, register_harness_builder, spec_for
+  - `bootstrap_lib/harnesses/renderers.py` (1,046줄) - 6개 하네스 render_*_agents + 6개 write_*_harness_files
+- 수정 파일:
+  - `bootstrap_workflow_kit.py` (2,468줄 → 53줄) - thin re-export entry
+- 삭제 파일:
+  - `bootstrap_harnesses/__init__.py` (154줄) - bootstrap_lib/harnesses/ 로 이동
+
+### 다음 세션 시작 포인트
+
+- TASK-V052-002 시작: `workflow_kit/pyproject.toml` 작성 (MCP SDK 1.27.0 / pydantic v2 호환)
+
+## TASK-V052-002 `workflow_kit` 정식 패키지 배포
+
+- 상태: planned
+- 우선순위: high
+- 요청일: 2026-06-06
+- 완료일:
+- 담당: yklee (Mavis / MiniMax Code orchestrator)
+- 호스트명: yklee-ui-MacBookAir.local
+- 호스트 IP: 192.168.0.139
+- 영향 문서: `workflow_kit/pyproject.toml` (신규), `workflow-source/pyproject.toml` (신규)
+
+## TASK-V052-003 외부 저장소 pilot validation 1건
+
+- 상태: planned
+- 우선순위: high
+- 요청일: 2026-06-06
+- 완료일:
+- 담당: yklee (Mavis / MiniMax Code orchestrator)
+- 호스트명: yklee-ui-MacBookAir.local
+- 호스트 IP: 192.168.0.139
+- 영향 문서: `workflow-source/examples/pilot_validation_<repo>.md` (신규)
