+++ ai-workflow/memory/release/v0.5.2/session_handoff.md
@@ -1,0 +1,61 @@
+# Session Handoff
+
+- 문서 목적: 현재 세션의 작업 상태와 다음 세션을 위한 인계 사항을 정리한다.
+- 범위: 현재 focus, 진행 중/차단/완료 작업, 핵심 변경, 다음 액션, 리스크
+- 대상 독자: AI 에이전트, 저장소 maintainer
+- 상태: in_progress
+- 최종 수정일: 2026-06-06
+- 관련 문서: [./state.json](./state.json), [../../work_backlog.md](../../work_backlog.md), [./PROJECT_PROFILE.md](./PROJECT_PROFILE.md)
+
+## Current Focus
+
+- **현재 브랜치**: `release/v0.5.2` (main #73f8f2f 에서 분기, 2026-06-06)
+- **현재 주 작업 축**: v0.5.2 = v0.5.1 의 plan 된 3개 후속 모두 실행
+  - TASK-V052-001: `bootstrap_workflow_kit.py` 1,855줄 → `bootstrap_lib/` 풀 리팩터
+  - TASK-V052-002: `workflow_kit` 정식 패키지 배포 (`pyproject.toml` 추가, `pip install -e .` 가능)
+  - TASK-V052-003: 외부 저장소 pilot validation 1건 (Devhub_example 추천)
+- **현재 기준선**: main 은 v0.5.1 (PR #13/#14/#15 머지) 까지 stable. v0.5.2 는 main 에서 분기, 0 commit. `ai-workflow/memory/release/v0.5.2/` 디렉터리 부트스트랩 진행 중
+- **메모리 layer 연속성**: v0.5.1 의 memory layer 가 같은 패턴으로 v0.5.2 에도 이식됨
+
+## Work Status
+
+- TASK-V052-001 `bootstrap_workflow_kit.py` 풀 리팩터 (1,855줄 → `bootstrap_lib/` 패키지화): done
+- TASK-V052-002 `workflow_kit` 정식 패키지 배포 (`pyproject.toml`): planned
+- TASK-V052-003 외부 저장소 pilot validation 1건: planned
+- (v0.5.1 의 TASK-V051-001..006 done — history 보존)
+
+## Key Changes
+
+- TASK-V052-001: `bootstrap_workflow_kit.py` 2,468줄 모놀리식 → `bootstrap_lib/` 6 모듈 패키지화
+  - `paths.py` (158줄) - Paths, HarnessDefinition, make_paths, *_path helpers
+  - `writes.py` (154줄) - write_text, copy_core_docs, build_manifest, rel
+  - `discovery.py` (163줄) - iter_repo_files, detect_package_scripts, guess_run_command, value_or_inferred, global_snippet_sources, IGNORED_DIRS
+  - `renderers.py` (351줄) - render_readme, render_project_profile, render_session_handoff, render_backlog_index, render_daily_backlog, render_assessment, load_template
+  - `mcp.py` (255줄) - MCP_SERVER_ALIAS, _mcp_server_command, _mcp_server_env, render_*_mcp_config × 5, write_mcp_config_files, MCP_CONFIG_RENDERERS
+  - `harnesses/__init__.py` (154줄) - 기존 registry 이동 (HARNESS_SPECS, HARNESS_FILE_BUILDERS, register_harness_builder, spec_for)
+  - `harnesses/renderers.py` (1,046줄) - 6개 하네스 render 함수 + write_*_harness_files dispatchers
+  - `__main__.py` (831줄) - parse_args, infer_project_context, selected_harnesses, update_dependencies, write_harness_files, main
+  - `bootstrap_workflow_kit.py` (53줄) - thin re-export entry (backward compat)
+  - 회귀: check_bootstrap.py 7/7 PASS (new/existing/opencode/gemini-cli/antigravity/minimax-code/enable_mcp), check_bootstrap_mcp_roundtrip.py 5/5 PASS
+  - v0.5.0 abort 했던 작업과 비교: 같은 풀 리팩터지만 "thin entry + 6 모듈" 단순화 + 매 모듈 옮길 때마다 회귀 확인
+
+## Next Actions
+
+- TASK-V052-002: `workflow_kit/pyproject.toml` + `workflow-source/pyproject.toml` 작성, `pip install -e .` 가능하게
+- TASK-V052-003: Devhub_example (또는 my_harness) 에 bootstrap 적용, 결과를 `workflow-source/examples/pilot_validation_<repo>.md` 로 기록
+- 각 TASK 완료 시 memory layer 갱신 (backlog TASK done 마크, session_handoff Work Status / Key Changes / Next Actions 갱신, state.json 재생성, linter status ok 확인)
+- 모두 완료 후 v0.5.1 + v0.5.2 통합 release 노트 + 태그 + release PR
+
+## Risks & Blockers
+
+- **리스크 #1 (HIGH)**: TASK-V052-001 풀 리팩터는 v0.5.0 에서 시도했다가 위험으로 abort. 이번엔 `bootstrap_harnesses/` registry 가 이미 추출되어 있어 그쪽이 더 성숙. 단, 1,855줄 → 패키지 분할 시 기존 `check_bootstrap.py` 의 모든 단언문이 그대로 통과해야 함 (회귀 위험)
+- **리스크 #2 (MED)**: TASK-V052-002 의 `pyproject.toml` 작성 시 `workflow-source` 디렉터리가 `workflow_kit` 패키지의 부모로 남아야 함 (MCP SDK 1.27.0 / pydantic v2 호환). `pip install -e workflow-source` 와 `pip install -e workflow_kit` 두 가지 모두 가능하게 구성
+- **리스크 #3 (MED)**: TASK-V052-003 pilot 의 결과가 v0.5.2 의 "성공" 기준이 됨. Devhub_example 에 적용 시 README/QUICKSTART 의 명령이 실제 외부 프로젝트에서도 동작해야 함
+- **제약**: Python 3.10+ (MCP SDK), `.venv-review/` 는 git 추적 금지, smoke test 회귀 0
+
+## 다음에 읽을 문서
+
+- [TASK-V052-001](./backlog/2026-06-06.md#task-v052-001)
+- [TASK-V052-002](./backlog/2026-06-06.md#task-v052-002)
+- [TASK-V052-003](./backlog/2026-06-06.md#task-v052-003)
+- [Maturity Matrix](../../../../workflow-source/core/maturity_matrix.json)
